refactor `SkipPath()` to utils
