+++ pkg/iac/scanners/terraform/parser/parser.go
@@ -11,10 +11,9 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/bmatcuk/doublestar/v4"
+	"github.com/aquasecurity/trivy/pkg/fanal/utils"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclparse"
-	"github.com/samber/lo"
 	"github.com/zclconf/go-cty/cty"
 
 	"github.com/aquasecurity/trivy/pkg/iac/ignore"
@@ -65,7 +64,6 @@ func New(moduleFS fs.FS, moduleSource string, opts ...Option) *Parser {
 		configsFS:      moduleFS,
 		logger:         log.WithPrefix("terraform parser").With("module", "root"),
 		tfvars:         make(map[string]cty.Value),
-		skipPaths:      []string{},
 	}
 
 	for _, option := range opts {
@@ -160,7 +158,7 @@ func (p *Parser) ParseFS(ctx context.Context, dir string) error {
 		if info.IsDir() {
 			continue
 		}
-		if SkipPath(realPath, CleanSkipPaths(p.skipPaths)) {
+		if utils.SkipPath(realPath, utils.CleanSkipPaths(p.skipPaths)) {
 			p.logger.Debug("Skipping path based on input glob", log.FilePath(realPath), log.Any("glob", p.skipPaths))
 			continue
 		}
@@ -366,26 +364,4 @@ func (s *paramParser) Parse(str string) bool {
 
 func (s *paramParser) Param() any {
 	return s.params
-}
-
-func CleanSkipPaths(skipPaths []string) []string {
-	return lo.Map(skipPaths, func(skipPath string, index int) string {
-		skipPath = filepath.ToSlash(filepath.Clean(skipPath))
-		return strings.TrimLeft(skipPath, "/")
-	})
-}
-
-func SkipPath(path string, skipPaths []string) bool {
-	path = strings.TrimLeft(path, "/")
-
-	// skip files
-	for _, pattern := range skipPaths {
-		match, err := doublestar.Match(pattern, path)
-		if err != nil {
-			return false // return early if bad pattern
-		} else if match {
-			return true
-		}
-	}
-	return false
 }
